💄 표 필터 ui 추가
클릭하면 밑에 보이기, 검색하기 버튼 추가
