+++ src/app/(main)/private-jobs/page.tsx
@@ -1,10 +1,11 @@
 import { Suspense } from 'react';
-import PriviteJobsPage from './PriviteJobsPage';
+import PriviteJobsPage from './PrivateJobsPage';
 
-export default function Page() {
+export default async function Page() {
   return (
     <Suspense fallback={<div>로딩 중...</div>}>
       <PriviteJobsPage />
     </Suspense>
   );
 }
+//여기서 데이터 페칭 받아서 넘겨주기
