+++ src/app/(main)/public-jobs/page.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from 'react';
 import PublicJobsPage from './PublicJobsPage';
 
-export default function Page() {
+export default async function Page() {
   return (
     <Suspense fallback={<div>로딩 중...</div>}>
       <PublicJobsPage />
