+++ src/constants/jobCategories.ts
